paas/application-manager/src/main/java/eu/ascetic/paas/applicationmanager/spreader/model/IaaSMessage.java
package eu.ascetic.paas.applicationmanager.spreader.model;

import javax.xml.bind.annotation.XmlAccessType;
import javax.xml.bind.annotation.XmlAccessorType;

/**
 * 
 * Copyright 2016 ATOS SPAIN S.A. 
 * 
 * Licensed under the Apache License, Version 2.0 (the License);
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 * http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * 
 * @author David Garcia Perez. Atos Research and Innovation, Atos SPAIN SA
 * e-mail david.garciaperez@atos.net 
 * 
 * Message for IaaS AMQP
 *
 *
 */

/*
 * Message format:
 * 
 * Topic: vm.489c0769-2ad3-4f76-a5d0-dedff0877f09.item.power
 * Message:
 * {"name":"power","value":16.9871,"units":"W","timestamp":1454434855}
 *
 */

@XmlAccessorType(XmlAccessType.FIELD)
public class IaaSMessage {
	public String name;
	public double value;
	public String units;
	public long timestamp;
	
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public double getValue() {
		return value;
	}
	public void setValue(double value) {
		this.value = value;
	}
	public String getUnits() {
		return units;
	}
	public void setUnits(String units) {
		this.units = units;
	}
	public long getTimestamp() {
		return timestamp;
	}
	public void setTimestamp(long timestamp) {
		this.timestamp = timestamp;
	}
	
	
}
